test: extract outbound delivery queue coverage
(cherry picked from commit fc140bb02bbd7dcbe1ea8a723ea3a117fdf8c924)
